move ocr score into module
