+++ examples/flowgrpo_trainer/run_qwen_image_ocr_lora_async_reward.sh
@@ -9,7 +9,7 @@ ocr_test_path=$WORKSPACE/data/ocr/test.parquet
 
 model_name=Qwen/Qwen-Image
 reward_model_name=Qwen/Qwen3-VL-8B-Instruct
-reward_function_path=examples/flowgrpo_trainer/reward_fn.py
+reward_function_path=verl_omni/utils/reward_score/genrm_ocr.py
 
 NUM_GPUS_ACTOR_ROLLOUT=4
 NUM_GPUS_REWARD=1
